Revert "fix: high CPU usage by replacing async context manager with setup/cle…"
This reverts commit 8aa021aee970fc01ccdb3d7f3242eee7fc3802cf.
